fix: change code to match changes in model from BE

+++ lib/models/incentive_model/incentive_model.dart
@@ -89,7 +89,7 @@ class IncentiveAccumulation {
   final double valueIncentivePrincipal;
 
   @JsonKey(name: 'achievement_percentage')
-  final int achievementPercentage;
+  final double achievementPercentage;
 
   @JsonKey(name: 'target_individual')
   final double? targetIndividual;
